Back to implementing `Sink` and `Stream` directly.
The previous approach of implementing the protocol directly as async
methods on `Connection` prevents concurrent sending and receiving.
Therefore we are back at implementing `Stream` and `Sink` directly with
an explicit state machine.

As most reading outside of the handshake phase is delegated to
tokio-codec, the `read_with_uninitialised_memory` feature is removed.