Use `Either` for extensions.
As a sort of custom copy-on-write type, we pass
`&mut Either<&[u8], BytesMut>` to `Extension::encode`. As soon as an
extension changed the data it will set this to `Either::Right` and from
then on, subsequent extensions may be able to re-use the `BytesMut`.

Also add `has_extensions` as an optimisation to `Sender` and `Receiver`,
so they never have to compete for the lock around extensions if there
are none.