Fix invalid continuation on mixed read/write
Add tests to verify proper handling of mixed receives and sends on the
same websocket.

When using the same websocket for reading and for writing using
`tokio::select!` the restart of the read will cause the read operation
to restart from the top of the stack of functions, this means that the
receive() operation must be restartable.

The existing implementation fills the Receiver.buffer with 2 bytes
when starting to read a header. The next restarted `receive()` call
will then see those two bytes (that haven't been read yet, just added
to the buffer) and interpret this incorrectly as two bytes that have
been read.

I don't understand why the tokio runtime / generated rust code doesn't
restart the future, and instead restarts from the top of the
async fn stack, but the small fix to read into a different buffer
makes header reads restartable. I think there might still be an issue
with reading huge WS messages which might suffer from the same issue,
but I can't produce such failures yet with the current test suite.