fix: ignore IO error after close has been sent
This PR fixes that after the close reason has been soketto is trying
to close down the I/O stream but it's possible that it's already closed
down by the other end after it receives the close message.

So, after this change the actual websocket close message is
prioritized and the actual I/O close status is ignored because
once the websocket close reason is sent to the other side then
it's regarded as clean shutdown.